+++ custom_components/maintenance_supporter/frontend/panel-chunks/object-dialog-5UQNDIIJ.js
@@ -1,2 +1,0 @@
-/*! maintenance_supporter frontend 2.84.0 */
-import{a}from"/maintenance_supporter_panelfiles/panel-chunks/chunk-SODCV2J2.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-ZX7ELJYQ.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-72EZBRXN.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-R6EFOWVO.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-UNDNGVCD.js";export{a as MaintenanceObjectDialog};